--- pages/404/404-server.js
@@ -1,9 +1,0 @@
-export default function get(req, res) {
-
-  res.status(404)
-  res.render(`404/404`, {
-    404:   true,
-    title: `Not Found`,
-  })
-
-}
