+++ pages/PageNotFound/PageNotFound-server.js
@@ -1,0 +1,15 @@
+export default function get(req, res) {
+
+  if (req.method !== `GET`) {
+    res.set(`Allow`, `GET`)
+    return res.sendStatus(405)
+  }
+
+  res.status(404)
+
+  res.render(`PageNotFound/PageNotFound`, {
+    404:   true,
+    title: `Page Not Found`,
+  })
+
+}
