+++ pages/PageNotFound/PageNotFound.hbs
@@ -1,10 +1,10 @@
 {{#section 'head'}}
-  <link rel=stylesheet href=/styles/404.css>
+  <link rel=stylesheet href=/styles/PageNotFound.css>
 {{/section}}
 
-<main id=main class=not-found-page>
+<main id=main class=page-not-found>
 
-  <h1 class=page-title>404: Not Found</h1>
+  <h1 class=page-title>404: Page Not Found</h1>
 
   <p>If you think you've reached this page in error, please <a class=link href='https://github.com/digitallinguistics/data-explorer/issues/new?{{ bugReportQuery }}'>open an issue on GitHub</a>.</p>
