+++ pages/PageNotFound/PageNotFound.less
@@ -1,7 +1,7 @@
 @import 'Link/Link.less';
 @import 'PageTitle/PageTitle.less';
 
-.not-found-page {
+.page-not-found {
 
   align-self     : center;
   max-inline-size: var(--measure);
