+++ pages/index.js
@@ -5,7 +5,7 @@ import Language        from './Language/Language-server.js'
 import LanguageLexemes from './LanguageLexemes/LanguageLexemes-server.js'
 import Languages       from './Languages/Languages-server.js'
 import Lexeme          from './Lexeme/Lexeme-server.js'
-import NotFound        from './404/404-server.js'
+import PageNotFound    from './PageNotFound/PageNotFound-server.js'
 import Project         from './Project/Project-server.js'
 import ProjectLexemes  from './ProjectLexemes/ProjectLexemes-server.js'
 import Projects        from './Projects/Projects-server.js'
@@ -20,7 +20,7 @@ export default {
   LanguageLexemes,
   Languages,
   Lexeme,
-  NotFound,
+  PageNotFound,
   Project,
   ProjectLexemes,
   Projects,
